Committing source code of the initial prototype of Isabelle running under Eclipse. The prototype is based on Isabelle/Scala libraries provided as part of the Isabelle distribution. Thus the projects require Scala-IDE for Eclipse to develop (2.0 branch works).
The UI allows launching an Isabelle system from Eclipse, provides theory editor with correct symbols and completion assistance; prover output; consoles, as well as several more things. It somewhat replicates what is available in Isabelle/JEdit system distributed with Isabelle 2011. One notable difference is that it is possible to only prove part of the document, while in Isabelle/JEdit, it is always the whole document that is being used.