# Working build with Isabelle2011-1 #
Added isabelle.scala.jar file, which helps with removing editor markers in dependent projects - not necessarily needed for build. If the editor error (red markers) persist, e.g. there are squiggles in the editor, but nothing in the Problems view, removing this jar file from the classpath, and adding it back again can help.