Change tree row height dynamically on Linux (issue #45)
On Windows and Mac there is a hack to change tree row height via reflection (see Eclipse bug 154341). Unfortunately, for Linux this is not possible. The official way of setting row height is via SWT.MeasureItem listener. Then trying to refresh the row height computation with Tree#clearAll(true). This does seem to work on Linux (tested on Ubuntu 12.10, 32-bit, GTK 2.24.13-0ubuntu2, I think), but I am not too sure about portability. Keeping the reflection-based solution for Windows and Mac.