Theories view (issue #43)
Theories view lists all open Isabelle theories. The theories are accompanied with a progress bar that provides an overview of their status in the prover. Furthermore, an icon provides an overall status of the theory (e.g. error, warning, finished or unfinished). When double-clicked, the corresponding theory file is opened.